% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeBlocksDistribution.R
\name{computeBlocksDistribution}
\alias{computeBlocksDistribution}
\title{computeBlocksDistribution}
\usage{
computeBlocksDistribution(x, blocks)
}
\arguments{
\item{x}{}

\item{blocks}{}
}
\description{
\code{computeBlocksDistribution}
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
