% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fill2DMean.R
\name{Fill2DMean}
\alias{Fill2DMean}
\title{Fill2DMean}
\usage{
Fill2DMean(GradTable, stat = "both", filter = 1, method = hmean,
  altGradTable = NA)
}
\arguments{
\item{GradTable}{is a filled summary of 2D simulations (usually output of \code{FillSetting})}

\item{stat}{'width' to compute mean and variance of widths,
'center' to compute mean and variance of centers and 'both' for both.}

\item{filter}{is the treshold of widths that will be filtered out (usually estimates < 1 are meaningless)}

\item{method}{to get mean (mean, hmean or any other funcion that tranform vector of numbers into one and takes rm.na = T argument)}

\item{altGradTable}{is a summary of the same 2D simulations with independent width/center estimates, these values are used in to fill failed estimates (< filter)}
}
\value{
The input GradTable with computed mean and variance of widths, centers or both
}
\description{
\code{Fill2DMean}
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
