% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fill2DSetting.R
\name{Fill2DSetting}
\alias{Fill2DSetting}
\title{Fill2DSetting}
\usage{
Fill2DSetting(sim, GradTable)
}
\arguments{
\item{sim}{A string of characters indicating a ReadSummary object. Please refer to documentation of \code{ReadSummary} function for detailed usage.}

\item{GradTable}{A string of characters indicating a ReadSetting object. Please refer to documentation of \code{ReadSetting} function for detailed usage.}
}
\value{
A data.frame object, integrating information about the settings used in the simulation run and the summary of the simulation run in a R-friendly format.
}
\description{
\code{Fill2DSetting} integrates the the output of \code{ReadSummary} and \code{ReadSetting} functions into a data.frame object. Warning: choose this function if Conjunction simulation was run in a 2D world; choose alternative function \code{FillSetting} instead if Conjunction simulation was run in a 1D world.
}
\examples{
{
   mysim <- ReadSummary(nameIn='../../Conjunction/out')
   myGradTable <- ReadSetting('../../Conjunction/setting.txt')
   myGradTable <- Fill2DSetting(sim=mysim, GradTable=myGradTable)
}

}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

