% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fill2DSettingByNLM.R
\name{Fill2DSettingByNLM}
\alias{Fill2DSettingByNLM}
\title{Fill2DSettingByNLM}
\usage{
Fill2DSettingByNLM(sim, GradTable)
}
\arguments{
\item{sim}{A string of characters indicating a ReadSummary object. Please refer to documentation of \code{ReadSummary} function for detailed usage.}

\item{GradTable}{A string of characters indicating a ReadSetting object. Please refer to documentation of \code{ReadSetting} function for detailed usage.}
}
\value{
A data.frame object, integrating information about the settings used in the simulation run and the summary of the simulation run in a R-friendly format.
}
\description{
\code{Fill2DSettingByNLM} integrates the the output of \code{ReadSummary} and \code{ReadSetting} functions into a data.frame object. Warning: choose this function if Conjunction simulation was run in a 2D world; choose alternative function \code{FillSetting} instead if Conjunction simulation was run in a 1D world.
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

