% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FillClosestS.R
\name{FillClosestS}
\alias{FillClosestS}
\title{FillClosestS}
\usage{
FillClosestS(multilocus, onelocus)
}
\arguments{
\item{multilocus}{a multilocus data frame filled by \code{FillSettingByHZAR}}

\item{onelocus}{a onelocus data frame filled by \code{FillSettingByHZAR}.
selections used for onelocus simulation should cover whole possible range (0.01 to 0.99)}
}
\value{
A data.frame object (multilocus) with new column ss;
}
\description{
\code{FillClosestS} is a function that finds effective selection (ss)
for each multilocus cline in the input data.frame.
ss is a selection acting on onelocus cline that is needed to maintaing cline of the same width.
}
\examples{
{
   # load data
   multilocus_sim = ReadSummary('multilocus_sim.out')
   ml_GradTable = ReadSetting('multilocus_setting.txt')
   ml_GradTable = FillSettingByHZAR(multilocus_sim, ml_GradTable)
   onelocus_sim = ReadSummary('onelocus_sim.out')
   onelocus_GradTable = ReadSetting('multilocus_setting.txt')
   onelocus_GradTable = FillSettingByHZAR(onelocus_sim, onelocus_GradTable)

   # compute ss
   myGradTable = FillClosestS(ml_GradTable, onelocus_GradTable)
}

}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

