% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FillSettingByHZAR.R
\name{FillSettingByHZAR}
\alias{FillSettingByHZAR}
\title{FillSettingByHZAR}
\usage{
FillSettingByHZAR(sim, GradTable, tails = "none")
}
\arguments{
\item{sim}{A string of characters indicating a ReadSummary object.
Please refer to documentation of \code{ReadSummary} function for detailed usage.}

\item{GradTable}{A string of characters indicating a ReadSetting object.
Please refer to documentation of \code{ReadSetting} function for detailed usage.}

\item{tails}{'none' fit logistic model,
'mirror' will fit a logistic function with exponenctial tails,
'auto' will choose a model with greater likelihood}
}
\value{
An input data.frame with, log likelihood of fitted model, estimates of width, center (tails = 'none'), deltaM and tauM (tail = 'auto' | 'mirror').
}
\description{
FillSettingByHZAR integrates the the output of \code{ReadSummary} and
\code{ReadSetting} functions into a data.frame object using package hzar
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
