% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterSetting.R
\name{FilterSetting}
\alias{FilterSetting}
\title{FilterSetting}
\usage{
FilterSetting(GradTable, par, treshold)
}
\arguments{
\item{GradTable}{}

\item{par}{name pattern of parameters to be filtered (ex. width)}

\item{treshold}{values to be replaced by NA (ex. 1)}
}
\description{
\code{FilterSetting} wilter out values lower than treshold of parameters starting by the name speciefied by par
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

