% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCompatibleSelectedLoci.R
\name{FindCompatibleSelectedLoci}
\alias{FindCompatibleSelectedLoci}
\title{FindCompatibleSelectedLoci}
\usage{
FindCompatibleSelectedLoci(loci, force_list = F)
}
\arguments{
\item{loci}{a vector of loci}

\item{force_list}{force list output [F] (see return)}
}
\value{
a vector (for one locus) or a list of vectors (for vector of loci) of compatible loci
each vector is sorted from the highest to lowest possible SL
ans since SL == L are always compatible, the first value of the vector corresponds to L.
}
\description{
compute a compatible selected loci to input number loci (a value or a vector)
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
