% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlockNumber.R
\name{getBlockNumber}
\alias{getBlockNumber}
\title{getBlockNumber}
\usage{
getBlockNumber(namePattern, save = 1, lim = 9)
}
\arguments{
\item{namePattern}{}

\item{save}{}

\item{lim}{}
}
\description{
\code{getBlockNumber}
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

