% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlockSummaries.R
\name{getBlockSummaries}
\alias{getBlockSummaries}
\title{getBlockSummaries}
\usage{
getBlockSummaries(path = ".", pattern = ".dat")
}
\arguments{
\item{path}{}
}
\value{
patterm
}
\description{
\code{getBlockSummaries} loads a type of output file 'blocks' from 1D or 2D
simulations
}
\examples{
{

}

}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
