% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoth.R
\name{getBoth}
\alias{getBoth}
\title{getBoth}
\usage{
getBoth(simtab, s)
}
\arguments{
\item{simtab}{}

\item{s}{}
}
\description{
\code{getBoth} is a function called inside \code{FillSetting}.
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

