% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCenter.R
\name{getCenter}
\alias{getCenter}
\title{getCenter}
\usage{
getCenter(simtab, GradTableLine)
}
\arguments{
\item{simtab}{}

\item{GradTableLine}{}
}
\description{
\code{getCenter}is a function called inside \code{Fill2DSetting} and \code{FillSetting}
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

