% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntagratedSelection.R
\name{getIntegratedSelection}
\alias{getIntegratedSelection}
\title{getIntegratedSelection}
\usage{
getIntegratedSelection(selection, beta)
}
\arguments{
\item{selection}{}

\item{beta}{}
}
\description{
\code{getIntegratedSelection} computes area under curve of selection function
defined by the Barton's selection function 1 - s (4x (1 - x)^beta)
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

