% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumberOfInvaders.R
\name{getNumberOfInvaders}
\alias{getNumberOfInvaders}
\title{getNumberOfInvaders}
\usage{
getNumberOfInvaders(sim, path = ".", pattern = ".dat")
}
\arguments{
\item{sim}{}

\item{path}{}
}
\value{
patterm
}
\description{
\code{getNumberOfInvaders}
}
\examples{
{
  
}

}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
