% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPj.R
\name{getPj}
\alias{getPj}
\title{getPj}
\usage{
getPj(namePattern, save = 1, lim = 9)
}
\arguments{
\item{namePattern}{}

\item{save}{}
}
\value{
lim
}
\description{
\code{getPj}
}
\examples{
{
  
}

}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

