% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopulationSize.R
\name{getPopulationSize}
\alias{getPopulationSize}
\title{getPopulationSize}
\usage{
getPopulationSize(LogFile, generation)
}
\arguments{
\item{LogFile}{}

\item{generation}{}
}
\description{
\code{getPopulationSize}
}
\examples{
{
  
}

}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

