% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSlope.R
\name{getSlope}
\alias{getSlope}
\title{getSlope}
\usage{
getSlope(simtab, GradTableLine)
}
\arguments{
\item{simtab}{}

\item{GradTableLine}{}
}
\description{
\code{getSlope} is a function called inside\code{FillSetting} and \code{Fill2DSetting}
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

