% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSstar.R
\name{getSstar}
\alias{getSstar}
\title{getSstar}
\usage{
getSstar(meanf, c, depth)
}
\arguments{
\item{meanf}{}

\item{c}{}

\item{depth}{}
}
\description{
\code{getSstar} is a function called inside\code{FillSetting}.
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
