% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVectorSplit.R
\name{getVectorSplit}
\alias{getVectorSplit}
\title{getVectorSplit}
\usage{
getVectorSplit(l)
}
\description{
\code{getVectorSplit} is a function called inside\code{ReadSetting} to integrate information from settings.txt file by parsing its content.
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
