% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWidthE.R
\name{getWidthE}
\alias{getWidthE}
\title{getWidthE}
\usage{
getWidthE(simtab)
}
\arguments{
\item{simtab}{is one of deta.frames from list of simulation summaries}
}
\description{
\code{getWidthE} is a function called inside \code{FillSetting}. It computes
an estimate of width based on Endler's definition of hybrid zone width
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

