% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmean.R
\name{hmean}
\alias{hmean}
\title{hmean}
\usage{
hmean(x, ...)
}
\arguments{
\item{x}{input vector of numbers}

\item{...}{just to avoid error when more parameters are passed (like na.rm)}
}
\description{
\code{hmean} computes an harmonic mean of all input non-NA values
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
