% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interceptToM.R
\name{interceptToM}
\alias{interceptToM}
\title{interceptToM}
\usage{
interceptToM(intercept, selection, theta)
}
\arguments{
\item{intercept}{}

\item{selection}{}

\item{theta}{}
}
\description{
\code{interceptToM}
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

