% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_logit.R
\name{inverse_logit}
\alias{inverse_logit}
\title{inverse_logit}
\usage{
inverse_logit(vector)
}
\description{
\code{inverse_logit} is function a reverse to \code{logit} transformation
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
