% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junctions2vector.R
\name{juncitons2vector}
\alias{juncitons2vector}
\title{juncitons2vector}
\usage{
juncitons2vector(ch)
}
\arguments{
\item{ch}{the string}
}
\value{
A vectors of loci
}
\description{
\code{juncitons2vector} converts a strings with blocks to vectors of ones one zeros (A and B loci)
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
