% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogBlocksDistribution.R
\name{LogBlocksDistribution}
\alias{LogBlocksDistribution}
\title{LogBlocksDistribution}
\usage{
LogBlocksDistribution(x, S, R, M)
}
\arguments{
\item{x}{}

\item{S}{}

\item{R}{}

\item{M}{}
}
\description{
\code{LogBlocksDistribution}
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

