% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot2DZone.R
\name{Plot2DZone}
\alias{Plot2DZone}
\title{Plot2DZone}
\usage{
Plot2DZone(GradTable, stat = "width", par1 = "s", par2 = "b", pal = NA,
  center = T)
}
\arguments{
\item{GradTable}{A dataframe of 2D simulation (usually output of \code{FillSetting})}

\item{stat}{The variable to be plot on x axis}

\item{par1}{The variable used for coloring}

\item{par2}{The variable used for line type}

\item{pal}{is a vector of colours used for colouring categories of par1.
By default the pallete will be generated as 'RdYlGn' color brewer pallete with apropriate number of categories}

\item{center}{T to centralize ploted stat for every simulation; F to plot stat without}
}
\description{
\code{PlotBoxplots} is a function for ploting a set of boxplots of a stat
of a set of simulations with respect to two parameters specified by user
}
\examples{
   Plot2DZone(GradTable, 'width', center = T)
   Plot2DZone(GradTable, 'center')


}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

