% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAverages.R
\name{PlotAverages}
\alias{PlotAverages}
\title{PlotAverages}
\usage{
PlotAverages(GradTable, stat, par1, par2, pal = NA, plot_sd = F,
  epsilon = 0.01, add = T, xlim = NA, ylim = NA, xlab = "",
  ylab = "", ...)
}
\arguments{
\item{GradTable}{A dataframe (usually output of \code{FillSetting}) or
\code{FillSettingByHZAR}}

\item{stat}{The variable to be plot on y axis}

\item{par1}{The variable to be plot on x axis}

\item{par2}{The variable used for the coloring of points}

\item{pal}{is a vector of colous used for colouring categories of the second parameter. By default the pallete will be generated as 'Set1' color brewer pallete with apropriate number of categories}

\item{plot_sd}{F; plot with averages also standard deviation}

\item{epsilon}{0.01 is width of the end of standard deviation bar}

\item{add}{T add points to existing plot, F create a new plot}

\item{xlab, }{ylab, xlim, ylim concernig creation of the new plot}

\item{...}{is passed to lines (not to ploints)}
}
\description{
\code{PlotAverages} add averages to a plot, usually combined with PlotStat.
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

