% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBLockDistr.R
\name{PlotBLockDistr}
\alias{PlotBLockDistr}
\title{plotBLockDistr}
\usage{
PlotBLockDistr(blocks, gradline)
}
\arguments{
\item{blocks}{}

\item{gradline}{}
}
\description{
\code{plotBLockDistr}
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

