% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBlocks.R
\name{PlotBlocks}
\alias{PlotBlocks}
\title{plotBLockDistr}
\usage{
PlotBlocks(blocks, onesim, gradline, horizontal = F)
}
\arguments{
\item{blocks}{is a nested list of blocks loaded using \code{ReadBlocks}}

\item{onesim}{is a data frame of stats of one simulation (element of list loaded by \code{ReadSummary})}

\item{gradline}{a line of corresponding simulation setting (line of data.frame loaded by \code{ReadSetting})}

\item{horizontal}{[F] plot space on horizontal axis and chromosome on vertical axis}
}
\description{
\code{plotBLockDistr}
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
