% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCF.R
\name{plotCF}
\alias{plotCF}
\title{plotCF}
\usage{
plotCF(FileName, GradTable, sim, i, ran = 5)
}
\arguments{
\item{FileName}{}

\item{GradTable}{}

\item{sim}{}

\item{i}{}

\item{ran}{}
}
\description{
\code{plotCF}
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

