% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHistograms.R
\name{PlotHistograms}
\alias{PlotHistograms}
\title{PlotHistograms}
\usage{
PlotHistograms(GradTable, stat, par1, par2, legend_position = NA, pal = NA,
  xlim = NA, ylim = NA)
}
\arguments{
\item{GradTable}{A dataframe (usually output of \code{FillSetting}) or
\code{FillSettingByHZAR}}

\item{stat}{The variable to be plot on y axis (used for construction of histograms)}

\item{par1}{The variable to be plot on x axis}

\item{par2}{The variable used for the coloring of histograms}

\item{legend_position}{is changing a position of the legend in the picture}

\item{pal}{is a vector of colous used for colouring categories of the second parameter. By default the pallete will be generated as 'Set1' color brewer pallete with apropriate number of categories}

\item{xlim}{is equvalent of parameter xlim from plot}

\item{ylim}{is equvalent of ylim from plot}
}
\description{
\code{PlotHistograms} is a function for ploting a set of histograms of a stat
of a set of simulations with respect to two parameters specified by user
}
\examples{
   PlotHistograms(GradTable, 'width_H', 's', 'D', 'topright')


}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

