% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLogitZone.R
\name{plotLogitZone}
\alias{plotLogitZone}
\title{plotLogitZone}
\usage{
plotLogitZone(FileName, GradTable, sim, sel, ran = 0, type = "pdf",
  leg = T)
}
\description{
\code{plotLogitZone}
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

