% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotParametricSpace.R
\name{PlotParametricSpace}
\alias{PlotParametricSpace}
\title{PlotParametricSpace}
\usage{
PlotParametricSpace(GradTable, global_margins = 0.1, denisity_haching = 10,
  cex = 2)
}
\arguments{
\item{GradTable}{}

\item{global_margins}{set margins around plot}

\item{denisity_haching}{set density of lines in rectangles}

\item{cex}{is size of all letters in plot (2 or 3 are reasonable)}
}
\description{
\code{PlotParametricSpace} plots range of selections, betas, lambdas and
number of chromosomes in one set of simulations
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
