% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotParametrixTemplate.R
\name{PlotParametricTemplate}
\alias{PlotParametricTemplate}
\title{PlotParametricTemplate}
\usage{
PlotParametricTemplate(xlim, x_by, cex = 2)
}
\description{
\code{PlotParametricTemplate} plots an empty plot for parameter
plotting function \code{PlotParametricSpace}
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
