% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSim.R
\name{PlotSim}
\alias{PlotSim}
\title{PlotSim}
\usage{
PlotSim(onesim, Gradline, legend_position = "bottomright", add = F,
  pal = NA, xlim = NA, ylim = NA)
}
\arguments{
\item{onesim}{is a data frame of stats of one simulation}

\item{legend_position}{is changing a position of the legend in the picture}

\item{add}{if True funciton will just add points to existing plot}

\item{pal}{is a vector of colous used for colouring categories of the second parameter. By default the pallete will be generated as 'Set1' color brewer pallete with apropriate number of categories}

\item{xlim}{is equvalent of parameter xlim from plot}

\item{ylim}{is equvalent of ylim from plot}

\item{GradTable}{A dataframe (usually output of \code{FillSetting}) or
\code{FillSettingByHZAR}}
}
\description{
\code{PlotSim} is a function for ploting an overview of one simulation
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

