% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotStat.R
\name{PlotStat}
\alias{PlotStat}
\title{PlotStat}
\usage{
PlotStat(GradTable, stat = "width_l", par1 = "s", par2 = NA,
  legend_position = "topright", add = F, pal = NA, xlim = NA,
  ylim = NA)
}
\arguments{
\item{GradTable}{A filled table of simulations}

\item{par1}{The variable to be plot on x axis}

\item{par2}{The variable used for the coloring of points}

\item{legend_position}{is changing a position of the legend in the picture}

\item{add}{if q= T funciton will just add points to existing plot}

\item{pal}{is a vector of colous used for colouring categories of the second parameter. By default the pallete will be generated as 'Set1' color brewer pallete with apropriate number of categories}

\item{xlim}{is equvalent of parameter xlim from plot}

\item{ylim}{is equvalent of ylim from plot}
}
\description{
\code{PlotStat} is a function for ploting a stat of set of simulation with respect to one or two parameters specified by user
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

