% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWC.R
\name{PlotWC}
\alias{PlotWC}
\title{plotWC}
\usage{
PlotWC(GradTable, sim, FileNam = NA)
}
\arguments{
\item{GradTable}{}

\item{FileName}{}

\item{SimTabName}{}
}
\description{
\code{plotWC}
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

