% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWidthEstimate.R
\name{plotWidthEstimate}
\alias{plotWidthEstimate}
\title{plotWidthEstimate}
\usage{
plotWidthEstimate(FileName, GradTable, sim, i, ran = 5)
}
\arguments{
\item{FileName}{}

\item{GradTable}{}

\item{sim}{}

\item{i}{}

\item{ran}{}
}
\description{
\code{plotWidthEstimate}
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

