% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWidthEstimates.R
\name{PlotWidthEstimates}
\alias{PlotWidthEstimates}
\title{PlotWidthEstimates}
\usage{
PlotWidthEstimates(GradLine, onesim, range = NA)
}
\arguments{
\item{GradLine}{a line from filled setting data.frame. Output of \code{FillSetting}.}

\item{onesim}{corresponding data.frame from list of simulation summaries}

\item{range}{defines how many demes around center should be plotted (NA for all)}
}
\description{
\code{PlotWidthEstimates} plots a mean hybrid indices with logit transformation
and estimates of width using Endler, fit of logistic curve and HZAR
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

