% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read0DSummary.R
\name{Read0DSummary}
\alias{Read0DSummary}
\title{Read0DSummary}
\usage{
Read0DSummary(summaryFile)
}
\arguments{
\item{summaryFile}{}
}
\value{
A list of data.frame objects, in which each list item, a data.frame, is a simulation run, and every row of the data.frame is a deme.
}
\description{
\code{Read0DSummary}
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

