% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadSetting.R
\name{ReadSetting}
\alias{ReadSetting}
\title{ReadSetting}
\usage{
ReadSetting(InputFile = "./setting.txt")
}
\arguments{
\item{Inputfile}{A string of characters indicating the settings file to be parsed. If no file is specified via InputFile, \code{ReadSetting} reads the file \code{settings.txt} in the current working directory.}
}
\value{
A data.frame object, featuring information about the settings used in the simulation run in a R-friendly format.
}
\description{
\code{ReadSetting} parses the settings file used for a simulation run using Conjunction software and converts it to a data.frame object.
}
\examples{
{
   myReadSetting = ReadSetting(InputFile='./mysetting.txt')
}

}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
