% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadSpatialSummary.R
\name{ReadSpatialSummary}
\alias{ReadSpatialSummary}
\title{ReadSpatialSummary}
\usage{
ReadSpatialSummary(summaryFile)
}
\arguments{
\item{summaryFile}{}
}
\value{
A list of data.frame objects, in which each list item, a data.frame, is a simulation run, and every row of the data.frame is a deme.
}
\description{
\code{ReadSpatialSummary}
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
