% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadSummary.R
\name{ReadSummary}
\alias{ReadSummary}
\title{ReadSummary}
\usage{
ReadSummary(nameIn)
}
\arguments{
\item{nameIn}{A string of characters indicating the simulation output file to be parsed. The output file can be defined following instructions on: https://github.com/KamilSJaron/Conjunction/wiki#usage}
}
\value{
A list of data.frame objects, in which each list item, a data.frame, is a simulation run, and every row of the data.frame is a deme.
}
\description{
\code{ReadSummary} parses the output file of a simulation run using Conjunction software and converts it to a list of data.frame objects.
}
\examples{
{
   myReadSummary = ReadSummary(nameIn='./mySimulationOutput')
}

}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

