% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{sigmoid}
\usage{
sigmoid(params, x)
}
\description{
\code{sigmoid} logistic function, parameters a, b, c are members of vector
params
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
