% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope2w.R
\name{slope2w}
\alias{slope2w}
\title{slope2w}
\usage{
slope2w(slope)
}
\arguments{
\item{slope}{}
}
\description{
\code{slope2w} width represented as parameter of logistic function have to be
multiplied by four to get a slope in hypertangent function. This function
computes width as an inverse of slope after conversion to hypertangent units
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
