% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeToTheta.R
\name{slopeToTheta}
\alias{slopeToTheta}
\title{slopeToTheta}
\usage{
slopeToTheta(slope)
}
\arguments{
\item{slope}{}
}
\description{
\code{slopeToTheta} function made for backward engenerring. It computes theta
from the slope of the log log distribution (which is linear)
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

