% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryToHZAR.R
\name{SummaryToHZAR}
\alias{SummaryToHZAR}
\title{SummaryToHZAR}
\usage{
SummaryToHZAR(onesim, gradline)
}
\arguments{
\item{onesim}{is a \code{data.frame}, member of list of summaries loaded by \code{ReadSummary}.
Package hzar has to be installed.}

\item{gradline}{is a line of setting table corresponding to the simulation}
}
\description{
\code{SummaryToHZAR} is a funciton converting list of summaries to HZAR objects
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

