% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trim2DSummary.R
\name{Trim2DSummary}
\alias{Trim2DSummary}
\title{Trim2DSummary}
\usage{
Trim2DSummary(sim, GradTable)
}
\arguments{
\item{sim}{A string of characters indicating a ReadSummary object. Please refer to documentation of \code{ReadSummary} function for detailed usage.}

\item{GradTable}{}
}
\value{
A data.frame object, integrating information about the settings used in the simulation run and the summary of the simulation run in a R-friendly format.
}
\description{
\code{Trim2DSummary} integrates the the output of \code{ReadSummary} and \code{ReadSetting} functions into a list of data.frame objects. Warning: choose this function if Conjunction simulation was run in a 2D world; choose alternative function \code{TrimSummary} instead if Conjunction simulation was run in a 1D world.
}
\examples{
{
   mySim=ReadSummary(nameIn='../../Conjunction/out')
   myReadSetting=ReadSetting(InputFile='../../Conjunction/setting.txt')
   myCenters=myReadSetting$C
   myTrim2DSummary = Trim2DSummary(sim=mySim, centers=myCenters)
}

}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

