% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrimSummary.R
\name{TrimSummary}
\alias{TrimSummary}
\title{TrimSummary}
\usage{
TrimSummary(sim, centers)
}
\arguments{
\item{sim}{A string of characters indicating a ReadSummary object. Please refer to documentation of \code{ReadSummary} function for detailed usage.}

\item{centers}{A string of characters indicating a center variable of ReadSetting object. Please refer to documentation of \code{ReadSetting} function for detailed usage.}
}
\value{
A data.frame object, integrating information about the settings used in the simulation run and the summary of the simulation run in a R-friendly format.
}
\description{
\code{TrimSummary} integrates the the output of \code{ReadSummary} and
\code{ReadSetting} functions into a list of data.frame objects.
Warning: choose this function if Conjunction simulation was run in a 1D world;
choose alternative function \code{Trim2DSummary} instead
if Conjunction simulation was run in a 2D world.
}
\examples{
{
   mySim=ReadSummary(nameIn='../../Conjunction/out')
   myReadSetting=ReadSetting(InputFile='../../Conjunction/setting.txt')
   myCenters=myReadSetting$C
   myTrimSummary = TrimSummary(sim=mySim, centers=myCenters)
}

}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

