% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twidth.R
\name{twidth}
\alias{twidth}
\title{twidth}
\usage{
twidth(s, sigma = sqrt(0.5))
}
\arguments{
\item{s}{selection}

\item{sigma}{dispersal}
}
\description{
\code{twidth} given selection and dispersal function returns
diffusion approximation of width of one locus cline
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

