% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlocksDistribution.R
\name{BlocksDistribution}
\alias{BlocksDistribution}
\title{BlocksDistribution}
\usage{
BlocksDistribution(x, S, R, M)
}
\arguments{
\item{x}{}

\item{S}{}

\item{R}{}

\item{M}{}
}
\description{
\code{BlocksDistribution} is a function called inside a plot function.
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

