% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitHZARmodel.R
\name{FitHZARmodel}
\alias{FitHZARmodel}
\title{FitHZARmodel}
\usage{
FitHZARmodel(AdaA, tails = "mirror")
}
\arguments{
\item{AdaA}{an hzar object}

\item{tails}{'none' for single locus clines}
}
\value{
A fitted hzar model
}
\description{
\code{FitHZARmodel} takes hzar object and use some arbitrarly chosen parameters to fit the model
}
\examples{
{
   AdaA <- SummaryToHZAR(sim[[1]], GradTable[1,])
   AdaAmodelData <- FitHZARmodel(AdaA);
}

}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

