% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitHZARmodel.R
\name{FitHZARmodel}
\alias{FitHZARmodel}
\title{FitHZARmodel}
\usage{
FitHZARmodel(mknAdaA)
}
\arguments{
\item{mknAdaA}{an hzar object}
}
\value{
A fitted hzar model
}
\description{
\code{FitHZARmodel} takes hzar object and use some arbitrarly chosen parameters to fit the model
}
\examples{
{
   mknAdaA <- SummaryToHZAR(sim[[1]], GradTable[1,])
   mknAdaAmodelData <- FitHZARmodel(mknAdaA);
}

}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

