% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAUFC.R
\name{GetAUFC}
\alias{GetAUFC}
\title{getAUFC}
\usage{
GetAUFC(selection, beta)
}
\arguments{
\item{selection}{}

\item{beta}{}
}
\description{
\code{getAUFC} computes Area Under Fitness Curve
defined by the Barton's fitness function 1 - s (4x (1 - x)^beta)
where, s is selection against the most admixed individuals and
beta is setrength of epistasic effect
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

