% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWidthE.R
\name{getWidthE}
\alias{getWidthE}
\title{getWidthE}
\usage{
getWidthE(simtab)
}
\arguments{
\item{simtab}{}
}
\description{
\code{getWidthE} is a function called inside \code{FillSetting}.
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

