% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBLockDistr.R
\name{plotBLockDistr}
\alias{plotBLockDistr}
\title{plotBLockDistr}
\usage{
plotBLockDistr(blocks, gradline)
}
\arguments{
\item{blocks}{}

\item{gradline}{}
}
\description{
\code{plotBLockDistr}
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

