% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWC.R
\name{plotWC}
\alias{plotWC}
\title{plotWC}
\usage{
plotWC(FileName, GradFileName, SimTabName)
}
\arguments{
\item{FileName}{}

\item{SimTabName}{}

\item{GradTable}{}
}
\description{
\code{plotWC}
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}

